feat(kumactl): better kumactl apply output (#13771)
## Motivation

The output of kumactl apply when there are multiple resources is missing
context. This is very poor UX, I've noticed this debugging e2e tests
where I had a bunch of yamls and just an error "something is wrong".

## Implementation information

- We now return errors for each entry in a multi resource yaml. This
makes it easier as all errors are returned at once
- We now log to stdout if the resource was updated or created.
- Added golden tests for outputs to make validating this behavior easier
- Avoid having golden files when the matching payload is empty (if no
file it means empty).

Fix #10099
Fix #12229

---------

Signed-off-by: Charly Molter <charly.molter@konghq.com>