docs(MADR): spiffe compliance - context and user stories (#13699)
## Motivation

As we begin work on making Kuma SPIFFE compliant, the first step is to
define what we aim to cover and what will be considered out of scope.
Additionally, I want to introduce the problem we're facing and outline
what we intend to solve.

xref: https://github.com/kumahq/kuma/issues/13661

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>
Co-authored-by: Krzysztof Słonka <slonka@users.noreply.github.com>