feat(meshcontext): move MeshServices, MeshExternalService and MeshMultizoneServices to new DestinationIndex inside BaseMeshContext (#13786)
## Motivation
For implementing new Inspect API and Dataplane Layout endpoint, we need
to compute list of outbounds. We have all the needed code in place, but
it was in `MeshContext`. We want to limit the scope of things that need
to be built for this endpoint. To achieve this, we can move this
information from `MeshContext` to `BaseMeshContext` as building it is
lightweight.

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>