ci(skaffold): simplify skaffold handling (#13819)
When added in #13698 we relied on modifying our tags and making things
work for skaffold Here we make skaffold work for us by making skaffold
use our tag format.

We also stop embedding the demo and have a `make dev/fetch-demo` target
instead

Finally use `docker/tag/kuma-cp` to remove the arch prefix from the
local tags. This is a good way to go about this because this is what's
already done in e2e tests:
https://github.com/kumahq/kuma/blob/8dc904d4a84bc8304784866501721e5153757b09/mk/k3d.mk#L219-L225

Signed-off-by: Charly Molter <charly.molter@konghq.com>