fix(dynconfig): change the direct response body size dynamically (#13853)
## Motivation

Since we introduced embedded DNS in `2.11.*`, the DNS data would be
stored in the direct response. However, the default max direct response
body size is 4 KiB, and if you exceed it, the envoy would return an
error. So, We need to change the data size.

## Implementation information

In the CP side, we calculate the size and set it in the XDS
configuration dynamically.


## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

Fix #13852 

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Icarus Wu <icaruswu66@qq.com>