fix(MeshTLS): skip MeshTLS policy xDS configurer when no policy is present
`MeshTLS` xDS configurer was running even when no `MeshTLS` policy was
applied, because it only checked for the presence of the policy type in
`proxy.Policies.Dynamic`. This resulted in Envoy resources being
configured without an actual policy.

This change adds a check for an empty `SingleItemRules.Rules` slice
to ensure that configuration is skipped when there are no `MeshTLS` policy
rules to apply.

Signed-off-by: Bart Smykla <bartek@smykla.com>