fix(MeshTLS): skip xDS configurer when no MeshTLS policy applies to the data plane (#13940)
## Motivation

The `MeshTLS` xDS configurer was applying configuration even when no
`MeshTLS` policy applied to the current data plane proxy. This happened
because the check only verified the presence of the policy type in
`proxy.Policies.Dynamic`, which is always set. As a result, Envoy
resources were generated even if no policy actually matched the data
plane proxy.

## Implementation information

The fix adds a check to ensure that the configurer exits early if there
are no `FromRules.InboundRules` or `GatewayRules.InboundRules`, meaning
no `MeshTLS` policy applies to the current data plane proxy. This
prevents unnecessary processing and avoids injecting unintended xDS
configuration.

## Supporting documentation

Fixes 13939

Signed-off-by: Bart Smykla <bartek@smykla.com>