ci(gha/pr-comments): allow `/format` comment to continue even if `make check` fails (#13948)
## Motivation

When someone comments `/format`, the job is supposed to fix formatting
and commit the changes. But after
[a53d0a6](https://github.com/kumahq/kuma/commit/a53d0a6f80ca46a8549964c16c30d9d4df6e8a7e),
it fails if `make check` finds changes. This means the job stops before
fixing anything, which is not what we want.

## Implementation information

This update adds `continue-on-error: true` to the step that runs `make
check`. Now, even if there are formatting issues, the job will keep
going and commit the changes.

Signed-off-by: Bart Smykla <bartek@smykla.com>