test(meshtls): improve meshtls test by running requets between asserts (#13952)
## Motivation

@lukidzi noticed a flake:
https://github.com/kumahq/kuma/actions/runs/16269540267/job/45939071797

```
• [FAILED] [40.527 seconds]
MeshTLS [It] should set cypher and version
github.com/kumahq/kuma/test/e2e_env/universal/meshtls/policy.go:332

Captured StdOut/StdErr Output

  [FAILED] Timed out after 30.001s.
  The function passed to Eventually failed at github.com/kumahq/kuma/test/e2e_env/universal/meshtls/policy.go:411 with:
  no stat found: &{[]}
Error: It 07/14/25 15:42:24.33
```

## Implementation information

Which we think was caused by the policy not being propagated quickly
enough so that the requests done in the previous lines did not trigger
the stats change. We moved everything after policy application to be in
eventually which should re-trigger stats if there is a failure in
eventually.

Signed-off-by: slonka <slonka@users.noreply.github.com>