refactor(MeshHTTPRoute/xds): simplify identifier fallback logic
Remove generic `IdentifierOrCallback` and `DefaultTo` helpers and replace
them with a simpler `IdentifierOrFallback` method on `DestinationService`.

This new helper resolves to a KRI identifier if unified naming parameter
is true and the outbound has an associated real resource. Otherwise, it
falls back to a provided string value.

Update `MeshHTTPRoute` listener generation to use the new method.

Signed-off-by: Bart Smykla <bartek@smykla.com>