feat(kuma-cp/xds): enable unified resource naming for MeshHTTPRoute listeners (#14038)
## Motivation

We want to support a new consistent way of naming Envoy resources, as
described in MADR-077. This makes it easier to understand, track, and
work with generated resources in tools and dashboards. The new naming is
optional and controlled by the `feature-unified-resource-naming` flag.

## Implementation information

We updated how outbound listeners are built for `MeshHTTPRoute` to
support the new naming format. The changes include:

- New helper functions to generate names in one place
(`destinationname.Resolve`, `MustResolve`)
- A simpler way to get names with fallback values
(`MaybeResolveKRIWithFallback`, `ResolveKRIWithFallback`,
`MaybeResolveKRI`)
- Updated `GenerateOutboundListener` to use the new naming when the
feature flag is on
- Renamed `DestinationService.ServiceName` to `KumaServiceTagValue` to
clarify that it comes from the `kuma.io/service` tag
- Renamed `MeshServiceElement` to `KumaServiceTagElement` for clarity
- Replaced `GetAddressOrDefault` with `GetAddressWithFallback` for
consistency
- Extracted names like `routeConfigName`, `virtualHostName`,
`listenerName`, and `statPrefix` using the new helpers
- Switched to a more flexible listener builder
- Added support for setting the `statPrefix` in listeners
- Cleaned up old logic and updated related tests

These changes let us switch between legacy and unified names without
breaking existing behavior.

## Supporting documentation

*
[MADR-077](https://github.com/kumahq/kuma/blob/52b8f5548739ecd1124181541f105bd6481b7ba5/docs/madr/decisions/077-migrating-to-consistent-and-well-defined-naming-for-non-system-envoy-resources-and-stats.md)
* Related issue: 13960

---------

Signed-off-by: Bart Smykla <bartek@smykla.com>