chore(kds): define user and internal errors and log user errors as info (#14083)
## Motivation

There are 2 reasons for NACK:
1. We've messed up with validation/feature-flags on Global. User applies
resource that's considered valid from Global's point of view and invalid
from Zone CP point of view. That shouldn't happen normally. **We'd like
to see these errors in logs (on error level)**.
2. On Zone CP user created resource with the name
`mesh-timeout-all-default-bv8z44b7bzb99xc7` and that's exactly the same
name as `mesh-timeout-all-default` will have when synced from Global to
Zone (original issue is https://github.com/kumahq/kuma/issues/10212). In
that case, the collision is created on user's side. It looks like we
can't do anything about it. We probably **shouldn't log this as an
error** in this case.

## Implementation information

Add "ErrUserNack" and "ErrInternalNack" and join the correct errors.

## Supporting documentation

xrel: https://github.com/kumahq/kuma/issues/10212

Signed-off-by: slonka <slonka@users.noreply.github.com>