ci(linter): fix linter not going through all properties (#14103)
## Motivation

There were bugs that caused not all fields to be checked by the linter:
1. Non leaf nodes because of the "continue" that happened after
recursion
2. Fields that were not in the same package were not checked (I thought
that the analyser loaded them if they were listed but it did not).

## Implementation information

For 1. we just remove the "continue" and add a check to not go back
above "default".
For 2. we preload the common packages so that we later can "jump" to
them using `CommonTypes`.

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

Fix https://github.com/kumahq/kuma/issues/14106

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>