chore(helm): switch kubectl image from bitnami to rancher
Bitnami will discontinue and remove the `bitnami/kubectl` image from
their registry in 2 weeks. While it is now available under the new
`bitnamisecure/kubectl` registry, it only contains new images, lacks
tags related to kubectl versions, and the publisher is not marked as
verified. This makes it unsuitable for backporting to release branches.

Switched to `rancher/kubectl` instead, keeping the same version
`1.33.3` with the corresponding image digest.

Signed-off-by: Bart Smykla <bartek@smykla.com>