chore(helm): switch kubectl image from bitnami to rancher (#14197)
## Motivation

Bitnami will discontinue and remove the `bitnami/kubectl` image from
their registry in 2 weeks. While kubectl is now available under the
`bitnamisecure/kubectl` registry, it only contains new images, lacks
tags corresponding to kubectl versions, and the publisher is not marked
as verified. This makes it unsuitable for backporting changes to
existing release branches.

To avoid disruption and ensure stability across versions, we are
switching to a different image source.

## Implementation information

Updated the Helm chart to use the `rancher/kubectl` image instead of
`bitnami/kubectl`, keeping the same version `1.33.3` and setting the
corresponding image digest. Rancher provides versioned and tagged
kubectl images, making it easier to align with our current and past
releases.

Signed-off-by: Bart Smykla <bartek@smykla.com>