chore(helm): switch kubectl image to rancher and bump patch
Bitnami will discontinue and remove the `bitnami/kubectl` image from
their registry in 2 weeks. While it is now available under the
`bitnamisecure/kubectl` registry, it only contains new images, has no
tags for kubectl versions, and the publisher is not marked as verified.
This makes it unsuitable for backports to release branches.

Switched to `rancher/kubectl` instead and bumped the patch version from
1.33.1 to the latest 1.33.3 with the corresponding image digest.

Signed-off-by: Bart Smykla <bartek@smykla.com>