chore(deps): update renovate config for openapi-tool
Updated `mk/docs.mk` to switch renovate metadata for `kumahq/openapi-tool`
from `github-tags` to `docker` datasource, using `semver-coerced` and
specifying the `ghcr.io` registry.

Added a Renovate package rule to disable pinning for `kumahq/openapi-tool`
in `renovate.json`. Pinning does not work for this dependency and always
results in failures, as it is managed by a custom regex manager for
Makefiles.

Signed-off-by: Bart Smykla <bartek@smykla.com>