ci(renovate): move security team assignment under vulnerabilityAlerts (#14495)
## Motivation

The `:assignAndReview(team:kuma-security-managers)` preset was
previously applied at the top level of the Renovate configuration. This
caused all Renovate PRs, including regular dependency updates, to be
assigned to the security managers team. The intention was to assign only
vulnerability alerts to the team.

## Implementation information

- Removed `:assignAndReview(team:kuma-security-managers)` from the
top-level `extends`
- Added `:assignAndReview(team:kuma-security-managers)` under the
`vulnerabilityAlerts` block so it only applies to security-related PRs

Signed-off-by: Bart Smykla <bartek@smykla.com>