fix(cni): should find & use cni conflist if confName is not set
When chained cni plugin is enabled and cni conf name is not provided,
the installerConfig PostProcess method searches for *.conf & *.conflist
extensions. If a conflist file is found, it validates it using conf file
checker func, resulting in validation error and using default kuma cni conf
name, which does not exist, leading to cni installer failure.

This fix chooses a approrpitate checker function based on the file extension.

Signed-off-by: AyushSenapati <a.p.senapati008@gmail.com>