ci(gha): automate golangci-lint version with Renovate (#14516)
## Motivation

The golangci-lint version in the workflow was hardcoded, which meant it
needed manual updates. This was easy to miss and left the workflow using
outdated versions. The goal was to make this version update automatic.

## Implementation information

- Introduced an environment variable `GOLANGCI_LINT_VERSION` annotated
for Renovate so it can track and update the version automatically
- Changed the golangci-lint job to use the environment variable instead
of a hardcoded version

## Result

The workflow now benefits from automated golangci-lint updates managed
by Renovate. Automated tracking and updates will start working once
https://github.com/Kong/public-shared-renovate/pull/56 is merged, a new
version of `Kong/public-shared-renovate` is released, and Renovate
updates the pinned preset version in this repository.

Signed-off-by: Bart Smykla <bartek@smykla.com>