chore(dev): remove ninja check from make install target (#14559)
## Motivation

We removed the need for ninja when we moved from `clang-format` to
`buf`. The install step still checked for `ninja`, which caused `make
install` to fail on machines without it even though it is not used
anymore. This updates the install target so it reflects the tools we
actually need.

## Implementation information

I removed the `cmd/check/ninja` dependency from the install phony target
in `mk/dev.mk`

Signed-off-by: Bart Smykla <bartek@smykla.com>