chore(generator): work around comments not correctly mapped (#14578)
## Motivation

In parent project for some reason only on CI the description is removed
when generating the OpenAPI schema.
Details in:
https://github.com/Kong/kong-mesh/issues/7376#issuecomment-3209289907

## Implementation information

Add if case for when "kuma" is parent project is in "./kuma" and not in
"../kuma".

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

closes
https://github.com/Kong/kong-mesh/issues/7376#issuecomment-3209289907

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>