fix(kuma-cp): configure Envoy internal addresses based on dp IPv6 support (#14652)
When a user is running kuma-dp on universal on machine that has IPv6
disabled, it won't be able to get xds config because Envoy will reject
`InternallAddressConfig` as it contains IPv6 by default. We want to skip
configuring IPv6 addresses if a machine on which kuma-dp does not have
support for it enabled

We are using transparent proxy helpers on dp to check if IPv6 support
was enabled, and we sent this information in DataplaneMetadata. Based on
this, we can later skip IPv6 internal addresses if needed. IPv6 support
will stay enabled by default to preserve backward compatibility.

Fix https://github.com/kumahq/kuma/issues/14643

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: Marcin Skalski <skalskimarcin33@gmail.com>