feat(oapi): add secret to openapi schema (#14745)
## Motivation

We want people to be able to define `secrets` in TF.

This is only part of the implem. We need global secrets as well. Let me
know if this should be a separate PR or in this one.

## Implementation information

Add `Secret` type to `ProtoTypeToType` struct. Make sure we collect
system types in oapi generation as well.

## Supporting documentation

xrel: https://github.com/kumahq/kuma/issues/14668

<!--
> Changelog: skip
-->
<!--
Uncomment the above section to explicitly set a [`> Changelog:` entry
here](https://github.com/kumahq/kuma/blob/master/CONTRIBUTING.md#submitting-a-patch)?
-->

---------

Signed-off-by: slonka <slonka@users.noreply.github.com>
Signed-off-by: Krzysztof Słonka <slonka@users.noreply.github.com>