fix(meshtrust): add trust independently of WorkloadIdentity (#14387)
## Motivation

We introduced MeshTrust, which allows users to provide a CA that the
mesh should use to accept traffic. Unfortunately, it was originally
tightly coupled to the existence of a WorkloadIdentity generated by the
MeshIdentity. This shouldn’t be the case, since MeshTrust should also
work without a MeshIdentity - for example, when operating with older
Mesh resources.

## Implementation information

*If Mesh.mTLS is enabled and MeshTrust is defined, we add the mesh mTLS
CA to the set of Trusts.
* A CAProvider is used in MeshContext to retrieve the mesh CA.
* By default, use the `system_trust_bundle` secret for the validation
context (when WorkloadIdentity is provided). Otherwise, use a
configurer.
* Sorting of Trusts was added in `dataplane_status_sink` to avoid
unnecessary updates.

## Supporting documentation

<!-- Is there a MADR? An Issue? A related PR? -->

Part of https://github.com/kumahq/kuma/issues/14340

---------

Signed-off-by: Lukasz Dziedziak <lukidzi@gmail.com>