test(framework): fix CNPG cluster naming in wait
The CNPG Helm chart creates clusters with name
`{primaryName}-cluster`, but the kubectl wait command was
looking for `{primaryName}` without the `-cluster` suffix.
This caused the wait to fail with "cluster not found" even
though the cluster was created successfully.

Also use the full resource name `clusters.postgresql.cnpg.io`
instead of the short form `cluster` for clarity.

Signed-off-by: Bart Smykla <bartek@smykla.com>